fixed as per review comments

+++ src/main/java/com/techm/orion/repositories/PortEntityRepository.java
@@ -13,10 +13,10 @@ import com.techm.orion.entitybeans.PortEntity;
 public interface PortEntityRepository extends JpaRepository<PortEntity, Long> { 
 	
 	@Query(value = "select count(port_name) from c3p_ports", nativeQuery = true)
-	float portNameCount();
+	int portNameCount();
 	
 	@Query(value = "select count(port_name) from c3p_ports where port_status =:portStatus", nativeQuery = true)
-	float statusPortNameCount(@Param("portStatus") String portStatus);
+	int statusPortNameCount(@Param("portStatus") String portStatus);
 
 	List<PortEntity> findBycardEntityCardId(int cardId);
